cleanup & fix replacement merge problems
