+++ designer/client/src/components/graph/node-modal/NodeSwitcherUtils.tsx
@@ -1,0 +1,160 @@
+import { cloneDeep, defaultsDeep, isArray, isPlainObject, mergeWith } from "lodash";
+import ProcessUtils from "../../../common/ProcessUtils";
+import { fakeComponentType, fakeNodeCreatorType } from "../../../reducers/selectors/getCreator";
+import { ComponentGroup, Edge, EdgeKind, NodeType, ProcessDefinitionData } from "../../../types";
+import NodeUtils from "../NodeUtils";
+import { ExpressionLang } from "./editors/expression/types";
+
+function mergeWithCustomizer<T>(object: T, source: T, path: string[] = []) {
+    return mergeWith(object, source, (val, src, key) => {
+        const fullPath = [...path, key];
+        switch (fullPath.join(".")) {
+            case "additionalFields.layoutData":
+                return src;
+            case "additionalFields.creatorType":
+                return val;
+            case "type":
+            case "service.id":
+            case "nodeType":
+                return val || null;
+            case "fields":
+            case "parameters":
+            case "service.parameters":
+                if (isArray(val)) {
+                    return val.map((parameter) =>
+                        mergeWithCustomizer(
+                            parameter,
+                            src.find((p) => p.name === parameter.name),
+                            [...fullPath, `[]`],
+                        ),
+                    );
+                }
+                return undefined;
+        }
+
+        if (isPlainObject(val) || isPlainObject(src)) {
+            return mergeWithCustomizer(val, src, fullPath);
+        }
+
+        return src;
+    });
+}
+
+function adjustEdges(outputEdges: Edge[], nextNode: NodeType, processDefinitionData: ProcessDefinitionData) {
+    switch (nextNode.type) {
+        case "Filter": {
+            let edgeKinds = [EdgeKind.filterTrue, EdgeKind.filterFalse];
+            return outputEdges.map((edge) => {
+                for (const kind of edgeKinds) {
+                    if (edge.edgeType?.type === kind) {
+                        edgeKinds = edgeKinds.filter((k) => edge.edgeType?.type !== k);
+                        return edge;
+                    }
+                }
+                return {
+                    ...edge,
+                    edgeType: {
+                        type: edgeKinds.shift(),
+                    },
+                };
+            });
+        }
+        case "Switch": {
+            return outputEdges.map((edge) => {
+                if ([EdgeKind.switchNext, EdgeKind.switchDefault].includes(edge.edgeType?.type)) {
+                    return edge;
+                }
+                return {
+                    ...edge,
+                    edgeType: {
+                        type: EdgeKind.switchNext,
+                        condition: {
+                            language: ExpressionLang.SpEL,
+                            expression: "true",
+                        },
+                    },
+                };
+            });
+        }
+        case "FragmentInput": {
+            let names = Object.keys(nextNode.ref?.outputVariableNames);
+            outputEdges
+                .filter(({ edgeType }) => {
+                    return edgeType?.type === EdgeKind.fragmentOutput && names.includes(edgeType?.name);
+                })
+                .forEach(({ edgeType }) => {
+                    names = names.filter((n) => n !== edgeType.name);
+                });
+
+            return outputEdges.map((edge) => {
+                if ([EdgeKind.fragmentOutput].includes(edge.edgeType?.type) && names.includes(edge.edgeType?.name)) {
+                    return edge;
+                }
+                return {
+                    ...edge,
+                    edgeType: {
+                        type: EdgeKind.fragmentOutput,
+                        name: names.shift(),
+                    },
+                };
+            });
+        }
+    }
+    if (NodeUtils.hasOutputs(nextNode, processDefinitionData)) {
+        return outputEdges.map(({ edgeType, ...edge }) => edge);
+    }
+    return [];
+}
+
+function compareNames(base: string, name: string): boolean {
+    return new RegExp(`^${base}( [0-9]+)?( \\(copy [0-9]+\\))*$`).test(name);
+}
+
+function isCustomName(editedNode: NodeType, componentGroups: ComponentGroup[]) {
+    const nodeCreatorType = fakeNodeCreatorType(editedNode);
+    const componentId = nodeCreatorType ? `${fakeComponentType}-${nodeCreatorType}` : ProcessUtils.determineComponentId(editedNode);
+    const components = componentGroups.flatMap((g) => g.components);
+    const component = components.find((c) => componentId === c.componentId);
+    return !compareNames(component.label, editedNode.$id || editedNode.id);
+}
+
+export function replaceNodeData(
+    editedNode: NodeType,
+    nextNodeData: NodeType,
+    processDefinitionData: ProcessDefinitionData,
+    outputEdges: Edge[] = [],
+    creatorType: string = null,
+    componentId?: string,
+) {
+    const object = cloneDeep(nextNodeData);
+
+    const componentGroups = processDefinitionData.componentGroups;
+
+    const isNameChanged = isCustomName(editedNode, componentGroups);
+    const nextNode = defaultsDeep(
+        {
+            id: isNameChanged || componentId ? editedNode.id : object.id,
+            $id: isNameChanged ? editedNode.id : componentId ? object.id : null,
+        },
+        mergeWithCustomizer(
+            {
+                ...object,
+                additionalFields: {
+                    ...object.additionalFields,
+                    creatorType: object.additionalFields?.creatorType || creatorType,
+                },
+            },
+            editedNode,
+        ),
+    );
+
+    const nextEdges = adjustEdges(
+        outputEdges.filter((e) => e.to),
+        nextNode,
+        processDefinitionData,
+    );
+    return {
+        nextNode,
+        nextEdges,
+    };
+}
